TfvcWorkspaceLocator: move workspace-related methods from CommandUtils
