+++ plugin/src/com/microsoft/alm/plugin/idea/tfvc/core/ChangelistBuilderStatusVisitor.java
@@ -134,7 +134,7 @@ class ChangelistBuilderStatusVisitor extends StatusProvider.StatusAdapter {
         if (lastChangeSet != null && !lastChangeSet.getChanges().isEmpty()) {
             final String serverPath = lastChangeSet.getChanges().get(0).getServerItem();
             final String originalPath = CommandUtils.getLocalPathSynchronously(serverContext, serverPath,
-                    CommandUtils.getWorkspaceName(project));
+                    TfvcWorkspaceLocator.getWorkspaceName(project));
 
             return TFSContentRevision.createRenameRevision(project,
                     VersionControlPath.getFilePath(originalPath, localPath.isDirectory()),
